+++ IsekaiMod/Content/Backgrounds/IsekaiBackgroundSelection.cs
@@ -1,4 +1,4 @@
-﻿using HarmonyLib;
+﻿using IsekaiMod.Utilities;
 using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
 using Kingmaker.Blueprints.Classes.Selection;
@@ -10,7 +10,6 @@ namespace IsekaiMod.Content.Backgrounds {
     internal class IsekaiBackgroundSelection {
 
         public static void Add() {
-            // Isekai Background Selection
             var IsekaiBackgroundSelection = Helpers.CreateBlueprint<BlueprintFeatureSelection>(IsekaiContext, "IsekaiBackgroundSelection", bp => {
                 bp.SetName(IsekaiContext, "Isekai");
                 bp.SetDescription(IsekaiContext, "Before you were hit by a truck, you were a...");
@@ -21,18 +20,12 @@ namespace IsekaiMod.Content.Backgrounds {
                 bp.m_AllFeatures = new BlueprintFeatureReference[0];
             });
 
-            // Add Isekai Background Selection to Background Selection
-            var BackgroundSelection = BlueprintTools.GetBlueprint<BlueprintFeatureSelection>("f926dabeee7f8a54db8f2010b323383c");
-            BackgroundSelection.m_AllFeatures = BackgroundSelection.m_AllFeatures.AddToArray(IsekaiBackgroundSelection.ToReference<BlueprintFeatureReference>());
+            StaticReferences.Selections.BackgroundSelection.AddToSelection(IsekaiBackgroundSelection);
         }
 
-        public static void Register(BlueprintFeature background) {
-            BlueprintFeatureSelection backgroundSelection = Get();
-            backgroundSelection.m_AllFeatures = backgroundSelection.m_AllFeatures.AddToArray(background.ToReference<BlueprintFeatureReference>());
-        }
-
-        public static BlueprintFeatureSelection Get() {
-            return BlueprintTools.GetModBlueprint<BlueprintFeatureSelection>(IsekaiContext, "IsekaiBackgroundSelection");
+        public static void AddToSelection(BlueprintFeature background) {
+            var backgroundSelection = BlueprintTools.GetModBlueprint<BlueprintFeatureSelection>(IsekaiContext, "IsekaiBackgroundSelection");
+            backgroundSelection.AddToSelection(background);
         }
     }
 }