+++ IsekaiMod/Content/Features/ExceptionalFeats/ExceptionalFeatSelection.cs
@@ -1,7 +1,5 @@
-﻿using HarmonyLib;
-using IsekaiMod.Utilities;
+﻿using IsekaiMod.Utilities;
 using Kingmaker.Blueprints;
-using Kingmaker.Blueprints.Classes;
 using Kingmaker.Blueprints.Classes.Selection;
 using Kingmaker.Designers.Mechanics.Recommendations;
 using TabletopTweaks.Core.Utilities;
@@ -15,6 +13,11 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats {
         private static readonly BlueprintFeatureSelection ExtraMythicAbilityMythicFeat = BlueprintTools.GetBlueprint<BlueprintFeatureSelection>("8a6a511c55e67d04db328cc49aaad2b8");
 
         public static void Add() {
+            // Add more Exceptional feats later
+            var ExceptionalFeatures = MythicFeatSelection.m_AllFeatures
+                    .RemoveFromArray(ExtraFeatMythicFeat.ToReference<BlueprintFeatureReference>())
+                    .RemoveFromArray(ExtraMythicAbilityMythicFeat.ToReference<BlueprintFeatureReference>());
+
             // The reason for two copies is to avoid a UI bug when exceptional feats are selected both in feat and bonus feat.
             var ExceptionalFeatSelection = Helpers.CreateBlueprint<BlueprintFeatureSelection>(IsekaiContext, "ExceptionalFeatSelection", bp => {
                 bp.SetName(IsekaiContext, "Exceptional Feats");
@@ -24,11 +27,7 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats {
                 bp.AddComponent<PureRecommendation>(c => {
                     c.Priority = RecommendationPriority.Good;
                 });
-
-                // Add more Exceptional feats later
-                bp.m_AllFeatures = MythicFeatSelection.m_AllFeatures
-                    .RemoveFromArray(ExtraFeatMythicFeat.ToReference<BlueprintFeatureReference>())
-                    .RemoveFromArray(ExtraMythicAbilityMythicFeat.ToReference<BlueprintFeatureReference>());
+                bp.m_AllFeatures = ExceptionalFeatures;
             });
             var ExceptionalFeatBonusSelection = Helpers.CreateBlueprint<BlueprintFeatureSelection>(IsekaiContext, "ExceptionalFeatBonusSelection", bp => {
                 bp.SetName(IsekaiContext, "Exceptional Feats");
@@ -38,30 +37,21 @@ namespace IsekaiMod.Content.Features.ExceptionalFeats {
                 bp.AddComponent<PureRecommendation>(c => {
                     c.Priority = RecommendationPriority.Good;
                 });
-
-                // Add more Exceptional feats later
-                bp.m_AllFeatures = MythicFeatSelection.m_AllFeatures
-                    .RemoveFromArray(ExtraFeatMythicFeat.ToReference<BlueprintFeatureReference>())
-                    .RemoveFromArray(ExtraMythicAbilityMythicFeat.ToReference<BlueprintFeatureReference>());
+                bp.m_AllFeatures = ExceptionalFeatures;
             });
 
-            // Add to Basic Feat Selection
-            var BasicFeatSelection = BlueprintTools.GetBlueprint<BlueprintFeatureSelection>("247a4068296e8be42890143f451b4b45");
-            BasicFeatSelection.m_AllFeatures = ThingsNotHandledByTTTCore.AddToFirst<BlueprintFeatureReference>(BasicFeatSelection.m_AllFeatures, ExceptionalFeatSelection.ToReference<BlueprintFeatureReference>());
-        }
-
-        public static void AddToSelection(BlueprintFeature feature) {
-            var ExceptionalFeatSelection = BlueprintTools.GetModBlueprint<BlueprintFeatureSelection>(IsekaiContext, "ExceptionalFeatSelection");
-            var ExceptionalFeatBonusSelection = BlueprintTools.GetModBlueprint<BlueprintFeatureSelection>(IsekaiContext, "ExceptionalFeatBonusSelection");
-            ExceptionalFeatSelection.m_AllFeatures = ExceptionalFeatSelection.m_AllFeatures.AddToArray(feature.ToReference<BlueprintFeatureReference>());
-            ExceptionalFeatBonusSelection.m_AllFeatures = ExceptionalFeatBonusSelection.m_AllFeatures.AddToArray(feature.ToReference<BlueprintFeatureReference>());
+            StaticReferences.Selections.BasicFeatSelection.AddToFirst(ExceptionalFeatSelection);
         }
 
         public static void AddToSelection(BlueprintFeatureSelection selection, BlueprintFeatureSelection bonusSelection) {
             var ExceptionalFeatSelection = BlueprintTools.GetModBlueprint<BlueprintFeatureSelection>(IsekaiContext, "ExceptionalFeatSelection");
             var ExceptionalFeatBonusSelection = BlueprintTools.GetModBlueprint<BlueprintFeatureSelection>(IsekaiContext, "ExceptionalFeatBonusSelection");
-            ExceptionalFeatSelection.m_AllFeatures = ExceptionalFeatSelection.m_AllFeatures.AddToArray(selection.ToReference<BlueprintFeatureReference>());
-            ExceptionalFeatBonusSelection.m_AllFeatures = ExceptionalFeatBonusSelection.m_AllFeatures.AddToArray(bonusSelection.ToReference<BlueprintFeatureReference>());
+            ExceptionalFeatSelection.AddToSelection(selection);
+            ExceptionalFeatBonusSelection.AddToSelection(bonusSelection);
+        }
+
+        public static BlueprintFeatureSelection Get() {
+            return BlueprintTools.GetModBlueprint<BlueprintFeatureSelection>(IsekaiContext, "ExceptionalFeatSelection");
         }
     }
 }