+++ IsekaiMod/Content/Features/IsekaiProtagonist/OverpoweredAbility/OverpoweredAbilitySelection.cs
@@ -1,19 +1,21 @@
 ﻿using HarmonyLib;
+using IsekaiMod.Content.Classes.IsekaiProtagonist;
+using IsekaiMod.Utilities;
 using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
 using Kingmaker.Blueprints.Classes.Prerequisites;
 using Kingmaker.Blueprints.Classes.Selection;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
 using TabletopTweaks.Core.Utilities;
+using UnityEngine;
 using static IsekaiMod.Main;
 
 namespace IsekaiMod.Content.Features.IsekaiProtagonist.OverpoweredAbility {
 
     internal class OverpoweredAbilitySelection {
-
+        private static readonly Sprite Icon_TrickFate = BlueprintTools.GetBlueprint<BlueprintAbility>("6e109d21da9e1c44fb772a9eca2cafdd").m_Icon;
         public static void Add() {
             // Overpowered Ability Selection
-            var Icon_TrickFate = BlueprintTools.GetBlueprint<BlueprintAbility>("6e109d21da9e1c44fb772a9eca2cafdd").m_Icon;
             var OverpoweredAbilitySelection = Helpers.CreateBlueprint<BlueprintFeatureSelection>(IsekaiContext, "OverpoweredAbilitySelection", bp => {
                 bp.SetName(IsekaiContext, "Overpowered Ability");
                 bp.SetDescription(IsekaiContext, "At 1st level, you get to select an Overpowered Ability.");
@@ -51,27 +53,33 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.OverpoweredAbility {
                 bp.m_Features = new BlueprintFeatureReference[0];
             });
 
-            // You can't select another Overpowered Ability from Mythic Abilities
-            OverpoweredAbilityMythicSelection.AddComponent<PrerequisiteNoFeature>(c => { c.m_Feature = OverpoweredAbilityMythicSelection.ToReference<BlueprintFeatureReference>(); });
+            // If MultipleMythicOPAbility setting is disabled, Mythic Overpowered Ability can only be selected once
+            if (!IsekaiContext.AddedContent.MultipleMythicOPAbility) {
+                OverpoweredAbilityMythicSelection.AddComponent<PrerequisiteNoFeature>(c => { c.m_Feature = OverpoweredAbilityMythicSelection.ToReference<BlueprintFeatureReference>(); });
+            }
 
-            // Add selection to mythic ability selection
-            var MythicAbilitySelection = BlueprintTools.GetBlueprint<BlueprintFeatureSelection>("ba0e5a900b775be4a99702f1ed08914d");
-            MythicAbilitySelection.m_AllFeatures = MythicAbilitySelection.m_AllFeatures.AppendToArray(OverpoweredAbilityMythicSelection.ToReference<BlueprintFeatureReference>());
+            // If RestrictMythicOPAbility is enabled, Mythic Overpowered Ability can only be selected by the Isekai Protagonist
+            if (IsekaiContext.AddedContent.RestrictMythicOPAbility) {
+                OverpoweredAbilityMythicSelection.AddPrerequisite<PrerequisiteClassLevel>(c => {
+                    c.m_CharacterClass = IsekaiProtagonistClass.GetReference();
+                    c.Level = 1;
+                });
+            }
+
+            StaticReferences.Selections.MythicAbilitySelection.AddToSelection(OverpoweredAbilityMythicSelection);
         }
 
         public static void AddToSelection(BlueprintFeature feature) {
-            var OverpoweredAbilitySelection = BlueprintTools.GetModBlueprint<BlueprintFeatureSelection>(IsekaiContext, "OverpoweredAbilitySelection");
-            var OverpoweredAbilitySelection2 = BlueprintTools.GetModBlueprint<BlueprintFeatureSelection>(IsekaiContext, "OverpoweredAbilitySelection2");
-            var OverpoweredAbilitySelectionVillain = BlueprintTools.GetModBlueprint<BlueprintFeatureSelection>(IsekaiContext, "OverpoweredAbilitySelectionVillain");
-            var OverpoweredAbilityMythicSelection = BlueprintTools.GetModBlueprint<BlueprintFeatureSelection>(IsekaiContext, "OverpoweredAbilityMythicSelection");
-            OverpoweredAbilitySelection.m_Features = OverpoweredAbilitySelection.m_Features.AddToArray(feature.ToReference<BlueprintFeatureReference>());
-            OverpoweredAbilitySelection.m_AllFeatures = OverpoweredAbilitySelection.m_AllFeatures.AddToArray(feature.ToReference<BlueprintFeatureReference>());
-            OverpoweredAbilitySelection2.m_Features = OverpoweredAbilitySelection2.m_Features.AddToArray(feature.ToReference<BlueprintFeatureReference>());
-            OverpoweredAbilitySelection2.m_AllFeatures = OverpoweredAbilitySelection2.m_AllFeatures.AddToArray(feature.ToReference<BlueprintFeatureReference>());
-            OverpoweredAbilitySelectionVillain.m_Features = OverpoweredAbilitySelectionVillain.m_Features.AddToArray(feature.ToReference<BlueprintFeatureReference>());
-            OverpoweredAbilitySelectionVillain.m_AllFeatures = OverpoweredAbilitySelectionVillain.m_AllFeatures.AddToArray(feature.ToReference<BlueprintFeatureReference>());
-            OverpoweredAbilityMythicSelection.m_Features = OverpoweredAbilityMythicSelection.m_Features.AddToArray(feature.ToReference<BlueprintFeatureReference>());
-            OverpoweredAbilityMythicSelection.m_AllFeatures = OverpoweredAbilityMythicSelection.m_AllFeatures.AddToArray(feature.ToReference<BlueprintFeatureReference>());
+            var OverpoweredAbilitySelections = new BlueprintFeatureSelection[] {
+                BlueprintTools.GetModBlueprint<BlueprintFeatureSelection>(IsekaiContext, "OverpoweredAbilitySelection"),
+                BlueprintTools.GetModBlueprint<BlueprintFeatureSelection>(IsekaiContext, "OverpoweredAbilitySelection2"),
+                BlueprintTools.GetModBlueprint<BlueprintFeatureSelection>(IsekaiContext, "OverpoweredAbilitySelectionVillain"),
+                BlueprintTools.GetModBlueprint<BlueprintFeatureSelection>(IsekaiContext, "OverpoweredAbilityMythicSelection"),
+            };
+            foreach(BlueprintFeatureSelection selection in OverpoweredAbilitySelections) {
+                selection.m_Features = selection.m_Features.AddToArray(feature.ToReference<BlueprintFeatureReference>());
+                selection.m_AllFeatures = selection.m_AllFeatures.AddToArray(feature.ToReference<BlueprintFeatureReference>());
+            }
         }
     }
 }